ETR-55 Feat: 플라스크 코드 수정에 따른 로직 개선
